move to `src`, fix by `pep8`, rename `id` to `_id`
